+++ push/sample/hms/src/main/java/com/pushpushgo/sdk/sample/push/hms/activity/MainActivity.kt
@@ -32,7 +32,7 @@ class MainActivity : AppCompatActivity() {
     findViewById<Button>(R.id.register).setOnClickListener {
       lifecycleScope.launch {
         try {
-          ppg.subscribeNow()
+          ppg.subscribe()
           Toast.makeText(this@MainActivity, "Subscribed!", Toast.LENGTH_SHORT).show()
         } catch (e: Exception) {
           Toast.makeText(this@MainActivity, "Can't subscribe! ${e.message}", Toast.LENGTH_SHORT).show()
@@ -42,7 +42,7 @@ class MainActivity : AppCompatActivity() {
 
     findViewById<Button>(R.id.unregister).setOnClickListener {
       lifecycleScope.launch {
-        ppg.unsubscribeNow()
+        ppg.unsubscribe()
       }
     }